update task adding, updating and others

+++ App/Controllers/TaskController.cs
@@ -6,6 +6,9 @@ using ConstructionManagementApp.App.Enums;
 using ConstructionManagementApp.App.Services;
 using ConstructionManagementApp.Events;
 using ConstructionManagementApp.App.Delegates;
+using ConstructionManagementApp.App.Exceptions;
+using ConstructionManagementApp.App.Models;
+
 
 
 namespace ConstructionManagementApp.App.Controllers
@@ -14,26 +17,42 @@ namespace ConstructionManagementApp.App.Controllers
     {
         private readonly TaskRepository _taskRepository;
         private readonly TaskAssignmentRepository _taskAssignmentRepository;
+        private readonly ProjectRepository _projectRepository;
         private readonly AuthenticationService _authenticationService;
+        private readonly UserRepository _userRepository;
+        private readonly RBACService _rbacService;
         public event LogEventHandler TaskAdded;
         public event LogEventHandler TaskDeleted;
         public event LogEventHandler TaskUpdated;
         public event LogEventHandler TaskCompleted;
+        public event LogEventHandler TaskStarted;
         public event LogEventHandler TaskAssigned;
         public event LogEventHandler TaskUnassigned;
 
 
-        public TaskController(TaskRepository taskRepository, TaskAssignmentRepository taskAssignmentRepository, AuthenticationService authenticationService)
+        public TaskController(TaskRepository taskRepository, TaskAssignmentRepository taskAssignmentRepository, AuthenticationService authenticationService, ProjectRepository projectRepository, RBACService rbacService, UserRepository userRepository)
         {
             _taskRepository = taskRepository;
             _taskAssignmentRepository = taskAssignmentRepository;
             _authenticationService = authenticationService;
+            _projectRepository = projectRepository;
+            _rbacService = rbacService;
+            _userRepository = userRepository;
         }
 
-        public void AddTask(string title, string description, TaskPriority priority, TaskProgress progress, int projectId)
+        public void AddTask(string title, string description, TaskPriority priority, TaskProgress progress, string projectName)
         {
+            
             try
             {
+                var project = _projectRepository.GetProjectByName(projectName);
+                if (project == null)
+                    throw new KeyNotFoundException($"Nie znaleziono projektu o nazwie {projectName}");
+                int projectId = project.Id;
+                if (!_rbacService.IsProjectManager(_authenticationService.CurrentSession.User, projectId, _projectRepository))
+                {
+                    throw new AuthorizationException("Brak uprawnień. Tylko menedżerowie tego projektu lub administratorzy mogą dodawać zadania.");
+                }
                 var task = new Task(title, description, priority, progress, projectId);
                 _taskRepository.AddTask(task);
                 Console.WriteLine("Zadanie zostało pomyślnie dodane.");
@@ -51,12 +70,20 @@ namespace ConstructionManagementApp.App.Controllers
 
         public void UpdateTask(int id, string title, string description, TaskPriority priority, TaskProgress progress)
         {
+            
             try
             {
                 var task = _taskRepository.GetTaskById(id);
                 if (task == null)
                     throw new KeyNotFoundException($"Nie znaleziono zadania o ID {id}.");
-
+                var project = _projectRepository.GetProjectById(task.ProjectId);
+                if (project == null)
+                    throw new KeyNotFoundException($"Nie znaleziono projektu o ID {id}");
+                int projectId = project.Id;
+                if (!_rbacService.IsProjectManager(_authenticationService.CurrentSession.User, projectId, _projectRepository))
+                {
+                    throw new AuthorizationException("Brak uprawnień. Tylko menedżerowie tego projektu lub administratorzy mogą dodawać zadania.");
+                }
                 task.Title = title;
                 task.Description = description;
                 task.Priority = priority;
@@ -87,6 +114,9 @@ namespace ConstructionManagementApp.App.Controllers
                 var task = _taskRepository.GetTaskById(id);
                 if (task == null)
                     throw new KeyNotFoundException($"Nie znaleziono zadania o ID {id}.");
+                var assignedUsers = _taskAssignmentRepository.GetWorkersAssignedToTask(id);
+                if (!assignedUsers.Any(worker => worker.Id == _authenticationService.CurrentSession.User.Id))
+                    throw new AuthorizationException("Brak uprawnień. Tylko przypisani użytkownicy mogą rozpocząć zadanie.");
                 task.Progress = TaskProgress.Completed;
                 _taskRepository.UpdateTask(task);
                 Console.WriteLine("Zadania zostało pomyślnie ukończone");
@@ -106,10 +136,50 @@ namespace ConstructionManagementApp.App.Controllers
             }
         }
 
+        public void StartTask(int id)
+        {
+            try
+            {
+                var task = _taskRepository.GetTaskById(id);
+                if (task == null)
+                    throw new KeyNotFoundException($"Nie znaleziono zadania o ID {id}");
+                var assignedUsers = _taskAssignmentRepository.GetWorkersAssignedToTask(id);
+                if (!assignedUsers.Any(worker => worker.Id == _authenticationService.CurrentSession.User.Id))
+                    throw new AuthorizationException("Brak uprawnień. Tylko przypisani użytkownicy mogą rozpocząć zadanie.");
+                task.Progress = TaskProgress.InProgress;
+                _taskRepository.UpdateTask(task);
+                Console.WriteLine("Zadanie zostało rozpoczęte");
+                TaskStarted?.Invoke(this, new LogEventArgs(_authenticationService.CurrentSession.User.Username, $"Zdanie o ID {id} zostało rozpoczęte"));
+            }
+            catch (KeyNotFoundException ex)
+            {
+                Console.WriteLine($"Błąd: {ex.Message}");
+            }
+            catch (ArgumentException ex)
+            {
+                Console.WriteLine($"Błąd: {ex.Message}");
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Nieoczekiwany błąd: {ex.Message}");
+            }
+        }
+
         public void DeleteTask(int id)
         {
             try
             {
+                var task = _taskRepository.GetTaskById(id);
+                if (task == null)
+                    throw new KeyNotFoundException($"Nie znaleziono zadania o ID {id}.");
+                var project = _projectRepository.GetProjectById(task.ProjectId);
+                if (project == null)
+                    throw new KeyNotFoundException($"Nie znaleziono projektu o ID {id}");
+                int projectId = project.Id;
+                if (!_rbacService.IsProjectManager(_authenticationService.CurrentSession.User, projectId, _projectRepository))
+                {
+                    throw new AuthorizationException("Brak uprawnień. Tylko menedżerowie tego projektu lub administratorzy mogą dodawać zadania.");
+                }
                 _taskRepository.DeleteTaskById(id);
                 Console.WriteLine("Zadanie zostało pomyślnie usunięte.");
                 TaskDeleted?.Invoke(this, new LogEventArgs(_authenticationService.CurrentSession.User.Username, $"Usunięto zadanie o ID {id}"));
@@ -128,12 +198,13 @@ namespace ConstructionManagementApp.App.Controllers
         {
             try
             {
-                var tasks = _taskRepository.GetAllTasks();
-                if (tasks.Count == 0)
-                {
-                    Console.WriteLine("Brak zadań w systemie.");
-                    return;
-                }
+                var userProjectIds = _rbacService.GetProjectsForUserOrManagedBy(_authenticationService.CurrentSession.User, _projectRepository);
+                if (!userProjectIds.Any())
+                    throw new InvalidOperationException("Nie jesteś przypisany do żadnych projektów.");
+                var tasks = _taskRepository.GetTasksByProjectIds(userProjectIds);
+
+                if (!tasks.Any())
+                    throw new InvalidOperationException("Nie jesteś przypisany do żadnych projektów.");
 
                 Console.WriteLine("--- Lista zadań ---");
                 foreach (var task in tasks)
@@ -151,6 +222,16 @@ namespace ConstructionManagementApp.App.Controllers
         {
             try
             {
+                var task = _taskRepository.GetTaskById(taskId);
+                if (task == null)
+                    throw new KeyNotFoundException($"Nie znaleziono zadania o ID {taskId}.");
+                if (!_rbacService.IsProjectManager(_authenticationService.CurrentSession.User, task.ProjectId, _projectRepository))
+                    throw new AuthorizationException("Brak uprawnień. Tylko manager projektu może przypisywać pracowników do zadania.");
+                var worker = _userRepository.GetUserByUsername(username);
+                if (worker == null)
+                    throw new KeyNotFoundException($"Nie znaleziono użytkownika o nazwie {username}.");
+                if (!_rbacService.IsWorkerInProjectTeam(worker, task.ProjectId, _projectRepository))
+                    throw new AuthorizationException("Brak uprawnień. Pracownik musi być członkiem zespołu przypisanego do projektu.");
                 _taskAssignmentRepository.AssignWorkerToTask(taskId, username);
                 Console.WriteLine($"Użytkownik o nazwie {username} został przypisany do zadania o ID {taskId}.");
                 TaskAssigned?.Invoke(this, new LogEventArgs(_authenticationService.CurrentSession.User.Username, $"Przypisano użytkownika {username} do zadania o ID {taskId}"));
@@ -165,9 +246,22 @@ namespace ConstructionManagementApp.App.Controllers
         {
             try
             {
+                var task = _taskRepository.GetTaskById(taskId);
+                if (task == null)
+                    throw new KeyNotFoundException($"Nie znaleziono zadania o ID {taskId}.");
+                if (!_rbacService.IsProjectManager(_authenticationService.CurrentSession.User, task.ProjectId, _projectRepository))
+                    throw new AuthorizationException("Brak uprawnień. Tylko manager projektu może usuwać pracowników z zadania.");
+                var worker = _userRepository.GetUserByUsername(username);
+                if (worker == null)
+                    throw new KeyNotFoundException($"Nie znaleziono użytkownika o nazwie {username}.");
+
                 _taskAssignmentRepository.RemoveWorkerFromTask(taskId, username);
                 Console.WriteLine($"Użytkownik o nazwie {username} został usunięty z zadania o Id {taskId}.");
                 TaskUnassigned?.Invoke(this, new LogEventArgs(_authenticationService.CurrentSession.User.Username, $"Usunięto użytkownika {username} z zadania o ID {taskId}"));
+            }
+            catch (KeyNotFoundException ex)
+            {
+                Console.WriteLine($"Błąd: {ex.Message}");
             }
             catch (Exception ex)
             {
@@ -196,5 +290,7 @@ namespace ConstructionManagementApp.App.Controllers
                 Console.WriteLine($"Błąd podczas pobierania użytkowników w zadania: {ex.Message}");
             }
         }
+
+        
     }
 }
